+++ rr/aaaa.js
@@ -7,8 +7,10 @@ const TINYDNS = require('../lib/tinydns')
 class AAAA extends RR {
   constructor (opts) {
     super(opts)
-    this.set('id', 28)
 
+    if (opts.tinyline) return this.fromTinydns(opts.tinyline)
+
+    this.set('id', 28)
     this.setAddress(opts?.address)
   }
 
@@ -20,19 +22,68 @@ class AAAA extends RR {
     this.set('address', val.toLowerCase()) // IETFs suggest only lower case
   }
 
+  getCompressed (f) {
+    this.compress(this.get(f))
+  }
+
   getRFCs () {
     return [ 3596 ]
   }
 
   /******  IMPORTERS   *******/
-  fromTinydns () {
-    // GENERIC =>  :fqdn:28:rdata:ttl:timestamp:lo
-    // AAAA       =>  3 fqdn : ip : x:ttl:timestamp:lo
-    // AAAA,PTR   =>  6 fqdn : ip : x:ttl:timestamp:lo
+  fromTinydns (str) {
+    let fqdn, ip, n, rdata, ttl, ts, loc
+
+    switch (str[0]) {
+      case ':':
+        // GENERIC  =>  :fqdn:28:rdata:ttl:timestamp:lo
+        [ fqdn, n, rdata, ttl, ts, loc ] = str.substring(1).split(':')
+        if (n != 28) throw new Error('SRV fromTinydns, invalid n')
+        // compressed format is needed for test comparisons
+        ip = this.compress(TINYDNS.octalToHex(rdata).match(/([0-9a-fA-F]{4})/g).join(':'))
+        break
+      case '3':
+      case '6':
+        // AAAA     =>  3 fqdn:rdata:x:ttl:timestamp:lo
+        // AAAA,PTR =>  6 fqdn:rdata:x:ttl:timestamp:lo
+        [ fqdn, rdata, ttl, ts, loc ] = str.substring(1).split(':')
+        ip = rdata.match(/(.{4})/g).join(':')
+        break
+    }
+
+    return new this.constructor({
+      type     : 'AAAA',
+      name     : fqdn,
+      address  : ip,
+      ttl      : parseInt(ttl, 10),
+      timestamp: ts,
+      location : loc !== '' && loc !== '\n' ? loc : '',
+    })
   }
 
   fromBind () {
     //
+  }
+
+  compress (val) {
+    return val
+      .replace(/\b(?:0+:){2,}/, ':')  // compress all zero
+      .split(':')
+      .map(o => o.replace(/\b0+/g, '')) // strip leading zero
+      .join(':')
+  }
+
+  expand (val, delimiter) {
+    if (delimiter === undefined) delimiter = ':'
+
+    const colons = val.match(/:/g)
+    if (colons && colons.length < 7) {
+      // console.log(`AAAA: restoring compressed colons`)
+      val = val.replace(/::/, ':'.repeat(9 - colons.length))
+    }
+
+    // restore compressed leading zeros
+    return val.split(':').map(s => s.padStart(4, 0)).join(delimiter)
   }
 
   /******  EXPORTERS   *******/
@@ -42,20 +93,8 @@ class AAAA extends RR {
   }
 
   toTinydns () {
-
-    let val = this.get('address')
-
-    const colons = val.match(/:/g)
-    if (colons && colons.length < 7) {
-      // console.log(`AAAA: restoring compressed colons`)
-      val = val.replace(/::/, ':'.repeat(9 - colons.length))
-    }
-
-    // restore compressed leading zeros
-    val = val.split(':').map(s => s.padStart(4, 0)).join('')
-
     // from AAAA notation (8 groups of 4 hex digits) to 16 escaped octals
-    const rdata = TINYDNS.packHex(val)
+    const rdata = TINYDNS.packHex(this.expand(this.get('address'), ''))
 
     return `:${this.get('name')}:28:${rdata}:${this.getEmpty('ttl')}:${this.getEmpty('timestamp')}:${this.getEmpty('location')}\n`
   }