+++ test/base.js
@@ -15,7 +15,6 @@ exports.valid = (type, validRecords) => {
       }
     })
   }
-
 }
 
 exports.invalid = (type, invalidRecords) => {
@@ -33,5 +32,33 @@ exports.invalid = (type, invalidRecords) => {
       throw new Error(`failed to throw`)
     })
   }
+}
 
+exports.toBind = (type, validRecords) => {
+
+  for (const val of validRecords) {
+    it(`exports ${val.type} to BIND (${val.name})`, async function () {
+      const r = new type(val).toBind()
+      if (process.env.DEBUG) console.dir(r)
+      assert.strictEqual(r, val.testR)
+    })
+  }
+}
+
+exports.toTinydns = (type, validRecords) => {
+
+  for (const val of validRecords) {
+    it(`exports ${val.type} to tinydns (${val.name})`, async function () {
+      const r = new type(val).toTinydns()
+      if (process.env.DEBUG) console.dir(r)
+      assert.strictEqual(r, val.testT)
+    })
+  }
+}
+
+exports.getRFCs = (type, valid) => {
+  it(`can retrieve RFCs for ${valid.type}`, async function () {
+    const r = new type(valid)
+    assert.ok(r.getRFCs().length)
+  })
 }