+++ test/caa.js
@@ -56,17 +56,18 @@ describe('CAA record', function () {
   base.valid(CAA, validRecords)
   base.invalid(CAA, invalidRecords)
 
-  for (const val of validRecords) {
-    it(`converts to BIND format ${val.name}`, async function () {
-      const r = new CAA(val).toBind()
-      if (process.env.DEBUG) console.dir(r)
-      assert.strictEqual(r, val.testR)
-    })
+  base.toBind(CAA, validRecords)
+  base.toTinydns(CAA, validRecords)
 
-    it(`converts to tinydns format ${val.name}`, async function () {
-      const r = new CAA(val).toTinydns()
+  base.getRFCs(CAA, validRecords[0])
+
+  for (const val of validRecords) {
+    it(`imports tinydns CAA (generic) record`, async function () {
+      const r = new CAA({ tinyline: val.testT })
       if (process.env.DEBUG) console.dir(r)
-      assert.strictEqual(r, val.testT)
+      for (const f of [ 'name', 'flags', 'tag', 'value', 'ttl' ]) {
+        assert.deepStrictEqual(r.get(f), val[f], `${f}: ${r.get(f)} !== ${val[f]}`)
+      }
     })
   }
 })