+++ test/uri.js
@@ -38,17 +38,18 @@ describe('URI record', function () {
   base.valid(URI, validRecords)
   base.invalid(URI, invalidRecords)
 
-  for (const val of validRecords) {
-    it(`converts to BIND format: ${val.name}`, async function () {
-      const r = new URI(val).toBind()
-      if (process.env.DEBUG) console.dir(r)
-      assert.strictEqual(r, val.testR)
-    })
+  base.toBind(URI, validRecords)
+  base.toTinydns(URI, validRecords)
 
-    it('converts to tinydns format', async function () {
-      const r = new URI(val).toTinydns()
+  base.getRFCs(URI, validRecords[0])
+
+  for (const val of validRecords) {
+    it(`imports tinydns URI (generic) record`, async function () {
+      const r = new URI({ tinyline: val.testT })
       if (process.env.DEBUG) console.dir(r)
-      assert.strictEqual(r, val.testT)
+      for (const f of [ 'name', 'priority', 'weight', 'target', 'ttl' ]) {
+        assert.deepStrictEqual(r.get(f), val[f], `${f}: ${r.get(f)} !== ${val[f]}`)
+      }
     })
   }
 })